Angular Agregando alretas en el momento de editar y borrar con sweet alert

+++ LearningAngular/src/app/components/article/article.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
 import { ArticleService } from '../../services/article.service';
 import { Article } from '../../models/article';
+import swal from 'sweetalert';
 
 import { Global } from '../../services/global';
 
@@ -45,16 +46,32 @@ export class ArticleComponent implements OnInit {
   }
 
   delete(id) {
-    this._articleService.delete(id).subscribe(
-      response => {
-        console.log(response);
-        this._router.navigate(['/blog']);
-      },
-      error => {
-        console.log(error);
-        this._router.navigate(['/blog']);
+    swal({
+      title: "¿Estás seguro?",
+      text: "Una vez borrado no podrás recuperarlo",
+      icon: "warning",
+      buttons: [true, true],
+      dangerMode: true,
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        swal("El artículo se ha elimidado correctamente", {
+          icon: "success",
+        });
+        this._articleService.delete(id).subscribe(
+          response => {
+            console.log(response);
+            this._router.navigate(['/blog']);
+          },
+          error => {
+            console.log(error);
+            this._router.navigate(['/blog']);
+          }
+        );
+      } else {
+        swal("Nada se ha borrado");
       }
-    );
+    });
   }
 
 }
